Add InfoModal as a component.
